grocery/requirements.txt
fastapi
uvicorn
pandas
requests
moya-ai
